pyproject.toml
[project]
name = "social-network"
version = "0.1.0"
description = ""
authors = [
    {name = "AnnyRomanova",email = "anna.romanova0818@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "authx (>=1.4.2,<2.0.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "python-jose (>=3.4.0,<4.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)"
]

[tool.poetry]
packages = [{include = "social_network", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
